refactor: add cancelableResourceController and test utils (#12292)
**Related Issue:** #7060

## Summary
This PR introduces the `useCancelable` controller, a utility designed to
manage `cancelable` resources (e.g., `debounced` and `throttled`
functions) within components. It ensures that these resources are
properly cleaned up when the component is disconnected from the DOM or
no longer needed. Additionally, it adds a test utility and a spec test
to ensure pending resources are canceled on disconnect.

---------

Co-authored-by: JC Franco <jfranco@esri.com>