docs(contributing): update blocked issue instructions (#12921)
**Related Issue:** n/a

## Summary
Updating the blocked issue instructions to reflect the change introduced
in this issue:
https://github.com/Esri/calcite-design-system/issues/12667

The reason for not including the "https://" in the full URL example is
so that Markdown doesn't auto hyperlink it. Apparently you can wrap
links in code blocks to avoid this but not sure if we wanted to do that
or not.