feat!: use the popover attribute to place floating components in the top-layer (#12904)
**Related Issue:** #10208 #13099

## Summary

- Updates components to use the `popover` attribute
- Deprecates the following css vars as they are no longer necessary
  - `--calcite-z-index-tooltip` 
  - `--calcite-z-index-popup`
  - `--calcite-z-index-dropdown`
- Removes the following internal css vars
  - `--calcite-floating-ui-z-index`
- Most of the updates are relatively simple but the changes to `alert`,
`dialog` and `sheet` are more complex because they have an internal
`embedded` property.
- Components that have an internal `embedded` property will still retain
styles for when they are slotted or embedded within other components.
- This is necessary as these components need to stay within their parent
component and will not be placed in the `top-layer`
  - This means the following CSS vars will remain.
    - `--calcite-z-index-modal`
    - `--calcite-z-index-overlay`
    - `--calcite-z-index-toast`'
- It also means we can't adjust the values for
`--calcite-z-index-header` and `--calcite-z-index-sticky` since the
above remain in embedded scenarios
- ~~These components using a `calcite-scrim` can use the `::backdrop`
pseudo element for stying when not `embedded`~~
- ~~In those cases above, a transparent div remains to handle clicking
on the backdrop to close the component~~

## Future thoughts

- It would be nice to test that components are indeed in the `top-layer`
within an e2e test

BREAKING CHANGE: The `alert`, `autocomplete`, `combobox`, `dialog`,
`dropdown`, `input-date-picker`, `popover`, `sheet`, and `tooltip`
components will now be placed within the document's
[top-layer](https://developer.mozilla.org/en-US/docs/Glossary/Top_layer)
when open. Setting `z-index` token values on them are no longer
necessary. Developers no longer need to specify the tokens
`--calcite-z-index-tooltip`, `--calcite-z-index-popup`,
`-calcite-z-index-dropdown`, or `--calcite-floating-ui-z-index`. Use
cases that may be affected would be custom `z-index` stacking and screen
shot tests with components that are initially open.