ci(actions): fix component label action (#13183)
**Related Issue:** n/a

## Summary
The component label action did not cover the case of no user response
and resulted in a weird "component" labels like `c-_no-response_`

In this PR:
- Make the "Which Component" field in the issue templates `required` but
set's a default value of "N/A" so it doesn't add any friction for folks.
- Adds "Which Component" dropdown to templates that it was previously
missing from (missed on initial pass)
- Add those new templates to the `updateIssueTemplateComponents` script
so that they update when components change.