fix(input-time-picker): always emit change event when user commits changes even when the value is directly set before or after the user commits the change (#13216)
**Related Issues:** #12391 #12889

## Summary

This PR fixes 2 issues: one where directly setting the value in a
listener responding to the `calciteInputTimePickerChange` event wasn't
always consistently working, and another where after directly setting
the value, a subsequent attempt at modifying the time value in the input
with the keyboard wasn't firing the `calciteInputTimePickerChange`
event. These 2 issues are related in that one directly sets the value
after the user has committed a change and the other changes the value
directly and then doesn't fire the change event when the user commits a
change. This update fixes both scenarios.