ci(actions): sync action changes to Monday (#13303)
**Related Issue:** #13239

## Summary
Ensures modifications to issues made through Actions are synced to
Monday.com.

This was already being done in a few scripts, but not all. The existing
script to handle the dispatched event is renamed and rewritten to
support syncing more types of data.

Unsure what happened with the renamed file, I used `git mv` but maybe
since too much content changed it just displays as removed/added.

Tested all but the 'needs more info' workflow in
https://github.com/ashetland/monday-testing.