ci(actions): fix regression label action (#13109)
**Related Issue:** n/a

## Summary
Currently the regression label is being added if the user enters
anything in the regression version at issue creation. This PR is to
update the action with a regex that will only add the label when a valid
semantic version is detected.

It also changes the name of the regression input in the issue templates
in hopes that it will convey to users that a regression version is
expected if they provide an input.