ci(actions): check if product mapping already exists (#13337)
**Related Issue:** #13182

## Summary
Improves the `createProductLabelIfNeeded` function and usage pattern by
using the GitHub label name for the map key instead of stripped name and
checking if the label aready exists.

- Stops the mapping from being created/updated multiple times due to it
being called in `createDropdownValue()` after it is already created.
- This also changes the function to return void, which makes the purpose
of the function clearer. It only creates the label mapping if needed,
and isn't relied on for returning the label name.
- Brings the mapping into alignment with the manual mappings by using
the GitHub label name as the key instead of a custom value (stripped
name).