fix(input-time-picker, time-picker): align time display closer to the browser's intended format (#13372)
**Related Issue:** #12723 

## Summary

This PR is a follow up to #13355 to add support for the `fr-CA` locale.
This update attempts to align the time format from the browser more
closely to the displayed time in both `input-time-picker` and
`time-picker` by removing relevant css spacing rules in favor of text
literals (that sometimes include whitespace depending on the locale)
returned by `Intl.DateTimeFormat`.

This PR also simplifies some internal code:

- Removes the internal utility `localizeTimeStringToParts` in favor of
`localizeTimeString({ parts: true })`.