docs(contributing): add issue refinement sub-section (#13430)
**Related Issue:** #12341

## Summary
Adds a sub-section under "Issue Management" documenting what steps to
take when an issue needs refinement. The steps are targeted for
developers/designers who are not sure how to proceed with issues that
lack information or have changed scope. PEs should also review this
section but will have to adapt the steps slightly. This approach was
decided as PEs will become more familiar with the process, while
developers/designers may only refer to this a handful of times.

---------

Co-authored-by: Kitty Hurley <khurley@esri.com>