docs: update keywords.json (#13395)
**Related Issue:** https://github.com/ArcGIS/icon-explorer/issues/6

## Summary
Updated the keywords.json file by merging codepoint values from
fantasticonrc.json.
Each icon entry now includes a new "codepoint" field, populated from the
codepoints list or set to an empty string if missing.
This is in the hope that we can update the reference page to Supports
improved search and integration for the icon font, enabling
codepoint-based lookups based on keywords.